unity-action@v1.0.1
- convert to TypeScript
