+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2024 Buildalon: Automate Unity
+Copyright (c) 2024 Virtual Maker Corporation
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
