+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "unity-action",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "A Github Action to execute Unity Editor command line arguments.",
   "author": "buildalon",
   "repository": {
@@ -24,10 +24,14 @@
     "@actions/io": "^1.1.3"
   },
   "devDependencies": {
-    "@vercel/ncc": "^0.34.0"
+    "@types/node": "^22.2.0",
+    "@vercel/ncc": "^0.34.0",
+    "shx": "^0.3.4",
+    "typescript": "^5.5.4"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "npm install && npm ci && ncc build src/index.js -o dist --source-map --license licenses.txt"
+    "build": "npm run clean && npm run bundle",
+    "bundle": "ncc build src/index.ts -o dist --source-map --license licenses.txt",
+    "clean": "npm install && shx rm -rf dist/ out/ node_modules/ && npm ci"
   }
 }
